refactor(mailbox): extraire send_queue_worker vers fichier dédié (SRP)
